generalize hiddev path for recent kernels
Hi,

I don't know if you are still actively maintaining the quickcalld
program, but it stopped working for me recently.  I tracked the problem
down to a change in the sysfs path used for the hiddev entry.  I've
attached a patch that fixed it for me, and hopefully does so in a way
that maintains compatibility across the sysfs layout change.  I'm
running Debian testing.  BTW, thanks a bunch for writing this tool!  It
saved me after a silly purchase :-).

                                                        -Kipp